Add --[no-]use-v1-api option to update command
Currently, the default API used by librarian is v1, with an option in
the install command to use the v3 API instead. This option is missing,
however, from the update command, which means that unofficial Forges
that implement the v3 API cannot be reliably updated from. This adds the
option to use either API version when updating.