Default to using API v3
This removes the hack to always use v3 for the official forge by
detecting the hostname. That was not taking puppet.com into account,
which is now the preferred domain.

This is also much better because there was no way to use a config file
to use API v3 by default. This is because librarian-puppet only allows
strings in the config and no normalization to false was possible.