Skip puppet2 tests
In 8d484314deded0c95af783b51898d687c28b42c9 a selection mechanism was
added and in c9fec1923f75c4864a6b63d0f24b15d4c4130895 the syntax was
changed. However, it doesn't appear to work anymore and now no tests are
executed. This statically skips the only test that can't be executed on
Puppet < 4. We don't test that anyway so nothing is lost.